fix: Add newlines to various printf format strings
The Dockerfile RUN to add the umask to /etc/bash.bashrc was not
including a newline so that subsequent appendings to the same file would
cause an error.  Conducted an audit of the repository for various
similar uses of printf and added newlines where it seemed to make sense.

Notable exceptions:
- apps/*/ci/latest.sh
- apps/*/scripts/sleep.sh

Signed-off-by: Dan Christensen <opello@opello.org>